requirements.txt
spotipy>=2.23.0
yt-dlp>=2023.11.16
python-dotenv>=1.0.0
requests>=2.31.0
mutagen>=1.47.0
Flask>=3.0.0
